Create Game Initializes player territories. Had to remove NotNull annotation due to db error. Change if resulting issues found.

+++ src/main/java/org/cost/territory/TerritoryController.java
@@ -114,7 +114,7 @@ class TerritoryController {
 
             PlayerTerritory playerTerritory = new PlayerTerritory();
             playerTerritory.setPlayerId(player.getPlayerId());
-            playerTerritory.setTerritoryId(territoryRequest.getTerritoryId());
+            playerTerritory.setTerritoryId((long) territoryRequest.getTerritoryId());
             player.getPlayerTerritoriesList().add(playerTerritory);
             playerRepository.save(player);
             return new ResponseEntity(HttpStatus.OK);
